requirements.txt
Flask==3.1.1
flask-cors==6.0.0
python-dotenv==1.1.0
google-generativeai==0.8.5
Pillow==11.2.1
Werkzeug==3.1.3
certifi==2025.4.26
requests==2.32.3
tqdm==4.67.1
